Adjustments. Remove branding. Make config file in /etc by default.

+++ freePbxConfig/voipConfigMaker.sh
@@ -1,6 +1,6 @@
 #! /bin/bash
 
-BASEDIR=/opt/mevoip/freePbxConfig 
+BASEDIR=/opt/voip/freePbxConfig 
 cd $BASEDIR/../
 svn update
 
@@ -11,11 +11,12 @@ if [ ! -e /usr/local/sbin/voipConfigMaker ]; then
 	ln -s $BASEDIR/voipConfigMaker.sh /usr/local/sbin/voipConfigMaker
 fi
 
-
-rm /opt/freepbx/html/admin/common/mainstyle.css
-rm /opt/freepbx/html/admin/images/freepbx_large.png
-ln -s $BASEDIR/webCustom/mainstyle.css /opt/freepbx/html/admin/common/mainstyle.css
-ln -s $BASEDIR/webCustom/freepbx_large.png /opt/freepbx/html/admin/images/freepbx_large.png
+if [ -e $BASEDIR/webCustom/mainstyle.css ]; then
+	rm /opt/freepbx/html/admin/common/mainstyle.css
+	rm /opt/freepbx/html/admin/images/freepbx_large.png
+	ln -s $BASEDIR/webCustom/mainstyle.css /opt/freepbx/html/admin/common/mainstyle.css
+	ln -s $BASEDIR/webCustom/freepbx_large.png /opt/freepbx/html/admin/images/freepbx_large.png
+fi
 
 
 cd $BASEDIR
